Add tests into main EngineTests class
